FEAT: Move OrderMap into the map submodule
For map/set symmetry, use map/set submodules. OrderMap itself is
reexported at the top, all other names are under ordermap::map::
